Mejora la organizacion y refactoriza la obtencion de personajes
