+++ src/App.jsx
@@ -1,0 +1,15 @@
+import './App.css'
+import 'bootstrap/dist/css/bootstrap.min.css'
+import Intro from './pages/intro/Intro'
+import IniciarSesion from './components/IniciarSesion'
+import Feed from './pages/feed/Feed'
+
+function App() {
+  return (
+    <>
+    <Feed />
+    </>
+  )
+}
+
+export default App
