--- src/components/app/App.jsx
@@ -1,13 +1,0 @@
-import './App.css'
-import 'bootstrap/dist/css/bootstrap.min.css'
-import Intro from '../intro/Intro'
-
-function App() {
-  return (
-    <>
-      <Intro />
-    </>
-  )
-}
-
-export default App
