+++ src/hooks/useSignWithG.jsx
@@ -1,6 +1,6 @@
 import { signInWithPopup, GoogleAuthProvider, linkWithPopup } from "firebase/auth";
-import { auth } from "../../firebase/firebase";
-import { useContextoUsuario } from "../../context/ContextoUsuario"
+import { auth } from "../firebase/firebase";
+import { useContextoUsuario } from "../context/ContextoUsuario"
 
 export function useSignWithG() {
     const { usuario, setUsuario } = useContextoUsuario();
