+++ src/main.jsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './components/app/App'
+import App from './App'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import './i18n'
 import { ContextoUsuarioProvider } from './context/contextoUsuario'
