+++ src/services/firebase/registroAnonimo.jsx
@@ -5,8 +5,8 @@ import {
 } from 'firebase/auth'
 import { auth } from './firebase'
 import { useContextoUsuario } from '../context/ContextoUsuario'
-import { useSignOut } from '../components/hooks/useSignOut'
-import { useSignWithG } from '../components/hooks/useSignWithG'
+import { useSignOut } from '../hooks/useSignOut'
+import { useSignWithG } from '../hooks/useSignWithG'
 
 export default function RegistroAnonimo() {
   const { cerrarSesion } = useSignOut()
